+++ lib/ts/recipe/thirdparty/recipeImplementation.ts
@@ -159,10 +159,18 @@ export default function getRecipeImplementation(recipeId: string, appInfo: Norma
                 config: input.config,
             });
 
-            const urlWithState = appendQueryParamsToURL(url, {
-                state,
-                redirect_uri: provider.getRedirectURL(),
-            });
+            // for some third party providers, the redirect_uri is set on the backend itself (for example in the case of apple). In these cases, we don't set them here...
+            const urlObj = new URL(url);
+            const alreadyContainsRedirectURI = urlObj.searchParams.get("redirect_uri") !== undefined;
+
+            const urlWithState = alreadyContainsRedirectURI
+                ? appendQueryParamsToURL(url, {
+                      state,
+                  })
+                : appendQueryParamsToURL(url, {
+                      state,
+                      redirect_uri: provider.getRedirectURL(),
+                  });
 
             // 4. Redirect to provider authorisation URL.
             redirectWithFullPageReload(urlWithState);