+++ lib/ts/recipe/emailpassword/types.ts
@@ -33,6 +33,7 @@ import {
     UserInput as AuthRecipeModuleUserInput,
     UserInputOverride as AuthRecipeUserInputOverride,
 } from "../authRecipeModule/types";
+import OverrideableBuilder from "supertokens-js-override";
 
 import { ComponentOverride } from "../../components/componentOverride/componentOverride";
 import { SignInHeader } from "./components/themes/signInAndUp/signInHeader";
@@ -63,7 +64,10 @@ export type UserInput = {
     signInAndUpFeature?: SignInAndUpFeatureUserInput;
     resetPasswordUsingTokenFeature?: ResetPasswordUsingTokenUserInput;
     override?: {
-        functions?: (originalImplementation: RecipeInterface) => RecipeInterface;
+        functions?: (
+            originalImplementation: RecipeInterface,
+            builder?: OverrideableBuilder<RecipeInterface>
+        ) => RecipeInterface;
         components?: ComponentOverrideMap;
     } & AuthRecipeUserInputOverride;
 } & AuthRecipeModuleUserInput<GetRedirectionURLContext, PreAPIHookContext, OnHandleEventContext>;
@@ -75,7 +79,10 @@ export type NormalisedConfig = {
     signInAndUpFeature: NormalisedSignInAndUpFeatureConfig;
     resetPasswordUsingTokenFeature: NormalisedResetPasswordUsingTokenFeatureConfig;
     override: {
-        functions: (originalImplementation: RecipeInterface) => RecipeInterface;
+        functions: (
+            originalImplementation: RecipeInterface,
+            builder?: OverrideableBuilder<RecipeInterface>
+        ) => RecipeInterface;
         components: ComponentOverrideMap;
     };
 } & NormalisedAuthRecipeModuleConfig<GetRedirectionURLContext, PreAPIHookContext, OnHandleEventContext>;
@@ -453,7 +460,7 @@ export type SignInAndUpState = {
     user?: User;
 };
 
-export interface RecipeInterface {
+export type RecipeInterface = {
     submitNewPassword: (input: {
         formFields: {
             id: string;
@@ -537,4 +544,4 @@ export interface RecipeInterface {
     >;
 
     doesEmailExist: (input: { email: string; config: NormalisedConfig }) => Promise<boolean>;
-}
+};