+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [0.17.2] - 2021-10-28
+
+### Changes
+
+-   Fixes issue two in https://github.com/supertokens/supertokens-node/issues/199
+-   Adds FDI 1.10 as supported
+
 ## [0.17.1] - 2021-10-28
 
 ### Changes